Refactor generation orchestrator factory into composables
