+++ app/frontend/src/composables/generation/index.ts
@@ -1,5 +1,6 @@
 export * from './useGenerationStudio';
 export * from './useGenerationStudioController';
+export * from './createGenerationOrchestrator';
 export * from './useGenerationTransport';
 export * from './useGenerationUI';
 export * from './useGenerationUpdates';
