+++ app/frontend/src/composables/generation/useGenerationUpdates.ts
@@ -8,7 +8,7 @@ import {
 import type {
   GenerationQueueClient,
   GenerationWebSocketManager,
-} from '@/services';
+} from '@/services/generation/updates';
 import type {
   GenerationJob,
   GenerationRequestPayload,
