+++ app/frontend/src/services/generation/index.ts
@@ -1,3 +1,2 @@
-export * from './orchestrator';
 export * from './updates';
 export * from './generationService';
